requirements.txt
colorama==0.4.4
gmail==0.6.3
PyAutoGUI==0.9.53
python-decouple==3.5
selenium==4.1.0
webdriver-manager==3.5.2
